support for iterate generator
